// src/api/models/createExperienceReqDTOHopeJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

export type CreateExperienceReqDTOHopeJob = typeof CreateExperienceReqDTOHopeJob[keyof typeof CreateExperienceReqDTOHopeJob];


export const CreateExperienceReqDTOHopeJob = {
  NONE: 'NONE',
  PLANNING: 'PLANNING',
  MARKETING: 'MARKETING',
  DESIGN: 'DESIGN',
  DEV: 'DEV',
  MEDIA: 'MEDIA',
  DATA: 'DATA',
} as const;